Allow XHR requests from remote domains
In the current form XHR requests are only allowed from localhost - for whatever reason.

In order to use the cover service from clientside applications on other domains we need to open up for other domains. This should open up for all domains.

I originally wanted to set the value to * but there seems to be problems with this approach when using the Nelmio bundle: https://github.com/nelmio/NelmioCorsBundle/issues/99. According to the comments . should work.